Original change: https://android-review.googlesource.com/c/platform/external/perfetto/+/1364717

+++ src/trace_processor/importers/ftrace/rss_stat_tracker.cc
@@ -62,7 +62,7 @@ void RssStatTracker::ParseRssStat(int64_t ts, uint32_t pid, ConstBytes blob) {
 
   if (utid) {
     context_->event_tracker->PushProcessCounterForThread(
-        ts, size, rss_members_[member], *utid);
+        ts, static_cast<double>(size), rss_members_[member], *utid);
   } else {
     context_->storage->IncrementStats(stats::rss_stat_unknown_thread_for_mm_id);
   }
