+++ src/trace_processor/importers/systrace/systrace_parser.cc
@@ -58,7 +58,7 @@ void SystraceParser::ParseZeroEvent(int64_t ts,
   systrace_utils::SystraceTracePoint point{};
   point.name = name;
   point.tgid = tgid;
-  point.value = value;
+  point.value = static_cast<double>(value);
 
   // The value of these constants can be found in the msm-google kernel.
   constexpr int32_t kSystraceEventBegin = 1 << 0;
@@ -98,7 +98,7 @@ void SystraceParser::ParseSdeTracingMarkWrite(int64_t ts,
   // the UI.
   point.tgid = 0;
 
-  point.value = value;
+  point.value = static_cast<double>(value);
   // Some versions of this trace point fill trace_type with one of (B/E/C),
   // others use the trace_begin boolean and only support begin/end events:
   if (trace_type == 0) {
